refactor: improve fetchOgpMetadata tests and structure
- Renamed the test suite for clarity, specifying it focuses on basic OGP fetching.
- Added afterEach hook to reset mocks after each test for better isolation.
- Removed outdated tests related to PDF and video URL handling, streamlining the test suite.
- Enhanced overall readability and organization of the test cases.
