+++ src/features/links/__tests__/utils/metadata.test.ts
@@ -17,11 +17,15 @@ jest.mock("link-preview-js", () => ({
   getLinkPreview: jest.fn(),
 }));
 
-describe("fetchOgpMetadata", () => {
+describe("fetchOgpMetadata - Basic OGP fetching", () => {
   beforeEach(() => {
     jest.clearAllMocks();
     // AsyncStorageはデフォルトでnullを返す（キャッシュなし）
     jest.mocked(AsyncStorage.getItem).mockResolvedValue(null);
+  });
+
+  afterEach(() => {
+    jest.resetAllMocks();
   });
 
   it("fetches and returns OGP metadata via link-preview-js", async () => {
@@ -137,108 +141,5 @@ describe("fetchOgpMetadata", () => {
 
     expect(result?.title).toBe("note ――つくる、つながる、とどける。");
     expect(result?.site_name).toBe("note（ノート）");
-  });
-
-  describe("Document and Media files", () => {
-    it("handles PDF URLs with proper fallback", async () => {
-      const result = await fetchOgpMetadata(
-        "https://example.com/documents/report-2024.pdf",
-      );
-
-      // PDFの場合はlink-preview-jsを呼ばずに直接フォールバック
-      expect(getLinkPreview).not.toHaveBeenCalled();
-      expect(result?.title).toBe("report-2024");
-      expect(result?.description).toBe("PDF Document");
-      expect(result?.image_url).toBeNull();
-    });
-
-    it("handles Japanese PDF filenames", async () => {
-      const result = await fetchOgpMetadata(
-        "https://example.co.jp/%E5%A0%B1%E5%91%8A%E6%9B%B8.pdf",
-      );
-
-      expect(result?.title).toBe("報告書");
-      expect(result?.description).toBe("PDF Document");
-    });
-
-    it("handles video URLs with proper fallback", async () => {
-      const result = await fetchOgpMetadata(
-        "https://example.com/videos/demo.mp4",
-      );
-
-      expect(getLinkPreview).not.toHaveBeenCalled();
-      expect(result?.title).toBe("demo");
-      expect(result?.description).toBe("Video");
-    });
-
-    it("handles PDF detected by Content-Type", async () => {
-      jest.mocked(getLinkPreview).mockResolvedValueOnce({
-        url: "https://example.com/document",
-        contentType: "application/pdf",
-        mediaType: "application",
-      } as unknown as Awaited<ReturnType<typeof getLinkPreview>>);
-
-      const result = await fetchOgpMetadata("https://example.com/document");
-
-      expect(result?.description).toBe("PDF Document");
-    });
-  });
-
-  describe("AsyncStorage cache", () => {
-    it("returns cached metadata when available", async () => {
-      const cachedMetadata = {
-        data: {
-          title: "Cached Title",
-          description: "Cached Description",
-          image_url: "https://example.com/cached.jpg",
-          site_name: "Example",
-          favicon_url: "https://example.com/favicon.ico",
-        },
-        timestamp: Date.now(),
-      };
-
-      jest
-        .mocked(AsyncStorage.getItem)
-        .mockResolvedValueOnce(JSON.stringify(cachedMetadata));
-
-      const result = await fetchOgpMetadata("https://example.com");
-
-      // キャッシュから取得されるため、link-preview-jsは呼ばれない
-      expect(getLinkPreview).not.toHaveBeenCalled();
-      expect(result).toEqual(cachedMetadata.data);
-    });
-
-    it("fetches from link-preview-js when cache is expired", async () => {
-      const expiredCache = {
-        data: {
-          title: "Expired Title",
-          description: null,
-          image_url: null,
-          site_name: "example.com",
-          favicon_url: null,
-        },
-        timestamp: Date.now() - 31 * 24 * 60 * 60 * 1000, // 31日前
-      };
-
-      jest
-        .mocked(AsyncStorage.getItem)
-        .mockResolvedValueOnce(JSON.stringify(expiredCache));
-
-      jest.mocked(getLinkPreview).mockResolvedValueOnce({
-        url: "https://example.com",
-        title: "Fresh Title",
-        description: "Fresh Description",
-        images: ["https://example.com/fresh.jpg"],
-        siteName: "Example",
-        mediaType: "website",
-      } as unknown as Awaited<ReturnType<typeof getLinkPreview>>);
-
-      const result = await fetchOgpMetadata("https://example.com");
-
-      // 期限切れキャッシュは削除され、link-preview-jsが呼ばれる
-      expect(AsyncStorage.removeItem).toHaveBeenCalled();
-      expect(getLinkPreview).toHaveBeenCalled();
-      expect(result?.title).toBe("Fresh Title");
-    });
   });
 });
